requirements.txt
#python-pptx==0.6.23
#Pillow==10.1.0
#pytesseract==0.3.10
#pypandoc-binary==1.12

# Backend
Flask
# IA
langchain
unstructured
python-docx
openai==0.28.0
python-dotenv
tiktoken
